Allow gitlab pages to be slow
Poorna keeps sending grey positives when gitlab pages is having operational issues. There's nothing we need to do about it and therefore, we're better off not triggering the warning. Ideally we need a way to ignore one failure and recheck in next cycle. But extending maxResponseTime to 120 seconds (from default of 30) will help in cases where the gitlab database is slow